Create Gardener and the Capybaras.c
Objective: given a string consisting of only letters "a" and "b", we're to find 3 strings a,b,c such that a<= b and c <= b or a >= b and c >= b. x <=y means that x must be earlier in the dictionary than y. 

Input: We're given a string consisting of letters "a" and "b"

Output: We're to output 3 strings a,b and c.

Observations:
- The string given is a+b+c
- Note the string only consist of a and b's, so if the first and last string has both a's just output it.
Questions: